requirements.txt
transformers>=4.48.0
datasets>=3.2.0
torch>=2.2.0
sentencepiece>=0.2.0
huggingface_hub>=0.28.0
pillow>=10.0.0
reportlab>=4.4.0
jupyter>=1.1.0
notebook>=7.0.0
